requirements.txt
asgiref==3.6.0
Django==4.2.1
django-author==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
setuptools-git==1.2
sqlparse==0.4.4
tzdata==2023.3
